Moving and Documenting
Documented new classes.

Moved 4 files to the io.javalin.core.compression package:

BrotliWrapper
GZIPWrapper
DynamicCompressionHandler
DynamicCompressionStrategy
