+++ src/main/java/io/javalin/core/compression/BrotliWrapper.java
@@ -1,4 +1,4 @@
-package io.javalin.core;
+package io.javalin.core.compression;
 
 import org.meteogroup.jbrotli.Brotli;
 import org.meteogroup.jbrotli.BrotliCompressor;
@@ -7,29 +7,39 @@ import java.io.Closeable;
 import java.io.IOException;
 import java.util.Arrays;
 
+
+/**
+ * This class acts as a middleman between JBrotli and Javalin.
+ */
 public class BrotliWrapper implements Closeable {
 
-    //DEFAULTS
     public static final Brotli.Mode DEFAULT_MODE = Brotli.DEFAULT_MODE;
     public static final int DEFAULT_LEVEL = 4;
     public static final int DEFAULT_LGWIN = Brotli.DEFAULT_LGWIN;
     public static final int DEFAULT_LGBLOCK = Brotli.DEFAULT_LGBLOCK;
 
-
-    //FIELDS
     private BrotliCompressor _brotliCompressor = new BrotliCompressor();
     private Brotli.Parameter _brotliParameter;
 
-
-    //CONSTRUCTORS
+    /**
+     * Default parameter
+     */
     public BrotliWrapper() {
         _brotliParameter = createParameter();
     }
 
+    /**
+     * Default parameter with custom compression level
+     *
+     * Valid level range is 0..11
+     */
     public BrotliWrapper(int level) {
         _brotliParameter = createParameter(level);
     }
 
+    /**
+     * Fully custom parameter
+     */
     public BrotliWrapper(Brotli.Mode mode, int level, int lgWin, int lgBlock) {
         _brotliParameter = createParameter(mode, level, lgWin, lgBlock);
     }
@@ -54,14 +64,17 @@ public class BrotliWrapper implements Closeable {
         return new Brotli.Parameter(mode, level, lgWin, lgBlock);
     }
 
+    /**
+     * @param input Byte array to compress
+     * @return Compressed byte array
+     */
     public byte[] compressByteArray(byte[] input) {
         byte[] output = new byte[input.length];
         int compressedLength = _brotliCompressor.compress(_brotliParameter, input, output);
         return Arrays.copyOfRange(output, 0, compressedLength);
     }
 
+    //Not used, but needs to be overriden because the class implements Closeable interface
     @Override
-    public void close() throws IOException {
-
-    }
+    public void close() throws IOException { }
 }
