+++ src/main/java/io/javalin/core/compression/GZIPWrapper.java
@@ -1,9 +1,13 @@
-package io.javalin.core;
+package io.javalin.core.compression;
 
 import java.io.IOException;
 import java.io.OutputStream;
 import java.util.zip.GZIPOutputStream;
 
+
+/**
+ * This class acts as a middleman between GZIP and Javalin.
+ */
 public class GZIPWrapper extends GZIPOutputStream {
 
     public GZIPWrapper(OutputStream out) throws IOException {
@@ -14,6 +18,9 @@ public class GZIPWrapper extends GZIPOutputStream {
         super(out, syncFlush);
     }
 
+    /**
+     * @param level GZIP compression level. Valid range is 0..9
+     */
     public GZIPWrapper setLevel(int level) {
         this.def.setLevel(level);
         return this;
