+++ src/test/kotlin/io/javalin/examples/TestServer.kt
@@ -1,7 +1,7 @@
 package io.javalin.examples
 
 import io.javalin.Javalin
-import io.javalin.core.DynamicCompressionStrategy
+import io.javalin.core.compression.DynamicCompressionStrategy
 
 private data class SillyObject(var fieldOne: String, var fieldTwo: String, var fieldThree: String)
 private fun getSomeObjects(numberOfObjects: Int) = (1..numberOfObjects).map { i -> SillyObject("f$i", "f$i", "f$i") }.toList()
